backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
playwright==1.54.0
python-multipart==0.0.6
python-dotenv==1.1.1
beautifulsoup4==4.13.4
requests==2.32.5
colorthief==0.2.1
webcolors==24.11.1
scikit-learn==1.7.1
numpy==2.3.2
cssselect==1.2.0
google-genai==1.31.0
python-dotenv==1.1.1
openai==1.101.0